feat: editable table tooltips
Implemented tooltips for error messages in the `EditableTable`
component.

These messages will be present at the top of a field that needs
correction, every time the field is focused or a key is pressed down
inside that field.

The `showErrors` method from the `EditableTable` component has been
deleted, since it is no longer in use.