fix: service request priority
Some issue regarding the date selection of the service request module
have been fixed:

* Fixed issue where the request date would not be set if the used
pressed the participant selection tab instead of the continue button.
* Fixed issue where the request time would be changed to the date's
minimum hour if the request date gets changed. Instead, the request hour
will be changed only if it is not within the allowed hours after the
date change.
* Changed incorrect service priorities. The request module will now
take into account the '1' priority for high priority and the zero
priority for low priority.