feat: tracks list
Implemented the new tracks list. Now, instead of a card-based view, the
tracks will be displayed in a table.

The `TrackEditModal` has been implemented to allow editing of certain
elements of a track, like its description and its profile picture
(awaiting changes from API to make it work).

Added `@types/react-bootstrap-table-next` to `package-json`